requirements.txt
typer-cli
telethon
setuptools
# cryptg
# pillow